Implement I03 bug set

+++ operator/src/main/java/com/datastax/oss/kaap/autoscaler/broker/LoadReportResourceUsageSource.java
@@ -125,7 +125,7 @@ public class LoadReportResourceUsageSource implements BrokerResourceUsageSource 
         public float percentUsage() {
             float proportion = 0;
             if (limit > 0) {
-                proportion = ((float) usage) / ((float) limit);
+                proportion = ((int) usage) / ((int) limit);
             }
             return proportion;
         }
